interpolation has been improved again :)

+++ src/data/player_data.hpp
@@ -30,7 +30,7 @@ struct SpecificIconData {
 };
 
 struct PlayerData {
-    uint64_t timestamp;
+    float timestamp;
     
     SpecificIconData player1;
     SpecificIconData player2;
